create a new behaviour to handle the weather client requests
