+++ .tool-versions
@@ -1,3 +1,3 @@
 elixir 1.10.3-otp-22
 erlang 22.3.3
-nodejs 14.2.0
+nodejs 12.13.1
