+++ lib/weather_application/weather_client.ex
@@ -1,0 +1,5 @@
+defmodule WeatherApplication.WeatherClient do
+  alias WeatherApplication.RequestParams
+
+  @callback request(params :: RequestParams.t()) :: String.t()
+end
