+++ lib/weather_application_web/live/page_live.ex
@@ -6,8 +6,6 @@ defmodule WeatherApplicationWeb.PageLive do
 
   alias WeatherApplication.RequestParams
 
-  @base_url "http://api.openweathermap.org/data/2.5/weather?"
-
   @impl true
   def mount(_params, _session, socket) do
     {:ok, assign(socket, zip: "", temperature: "", units: "imperial", results: %{})}
@@ -15,7 +13,7 @@ defmodule WeatherApplicationWeb.PageLive do
 
   @impl true
   def handle_event("search", %{"zip" => zip}, %{assigns: %{units: units}} = socket) do
-    temperature = parse_weather(%RequestParams{zip: zip, units: units})
+    temperature = client().request(%RequestParams{zip: zip, units: units})
 
     {:noreply, assign(socket, zip: zip, temperature: temperature)}
   end
@@ -25,33 +23,5 @@ defmodule WeatherApplicationWeb.PageLive do
     {:noreply, assign(socket, :units, units)}
   end
 
-  @spec parse_weather(map()) :: String.t()
-  defp parse_weather(%RequestParams{zip: zip} = params) do
-    case zip do
-      "" ->
-        ""
-
-      _zip ->
-        url = create_url(params)
-        {:ok, {{'HTTP/1.1', 200, 'OK'}, _headers, body}} = :httpc.request(:get, {url, []}, [], [])
-
-        {:ok, %{"main" => %{"temp" => temp}}} =
-          body
-          |> List.to_string()
-          |> Jason.decode()
-
-        temp
-    end
-  end
-
-  @spec create_url(RequestParams.t()) :: charlist()
-  defp create_url(%RequestParams{zip: zip, units: units}) do
-    String.to_charlist("#{base_url()}zip=#{zip}&units=#{units}&appid=#{api_key()}")
-  end
-
-  @spec base_url() :: String.t()
-  defp base_url, do: @base_url
-
-  @spec api_key() :: String.t()
-  defp api_key, do: Application.get_env(:weather_application, :openweather_api_key)
+  defp client, do: Application.get_env(:weather_application, :weather_client)
 end
