+++ phoenix_static_buildpack.config
@@ -1,1 +1,1 @@
-node_version=14.2.0
+node_version=12.13.1
